+++ bin/lhm-spec-clobber.sh
@@ -16,8 +16,6 @@ lhmkill() {
 echo stopping homebrew running mysql instance
 ls -lrt -d -1 ~/Library/LaunchAgents/* |  grep 'mysql.plist' | xargs launchctl unload -w
 
-lhmkill
-
 echo removing $basedir
 rm -rf "$basedir"