+++ spec/unit/table_spec.rb
@@ -16,19 +16,26 @@ describe Lhm::Table do
   end
 
   describe 'constraints' do
+    def set_columns(table, columns)
+      table.instance_variable_set('@columns', columns)
+    end
+
     it 'should be satisfied with a single column primary key called id' do
       @table = Lhm::Table.new('table', 'id')
-      @table.satisfies_primary_key?.must_equal true
+      set_columns(@table, { 'id' => { :type => 'int(1)' } })
+      @table.satisfies_id_column_requirement?.must_equal true
     end
 
-    it 'should not be satisfied with a primary key unless called id' do
+    it 'should be satisfied with a primary key not called id, as long as there is still an id' do
       @table = Lhm::Table.new('table', 'uuid')
-      @table.satisfies_primary_key?.must_equal false
+      set_columns(@table, { 'id' => { :type => 'int(1)' } })
+      @table.satisfies_id_column_requirement?.must_equal true
     end
 
-    it 'should not be satisfied with multicolumn primary key' do
-      @table = Lhm::Table.new('table', ['id', 'secondary'])
-      @table.satisfies_primary_key?.must_equal false
+    it 'should not be satisfied if id is not numeric' do
+      @table = Lhm::Table.new('table', 'id')
+      set_columns(@table, { 'id' => { :type => 'varchar(255)' } })
+      @table.satisfies_id_column_requirement?.must_equal false
     end
   end
 end